+++ CHANGELOG.md
@@ -13,7 +13,7 @@ DEPS_CLJ_TOOLS_VERSION=1.11.1.1165 bb clojure
 ## Unreleased
 
 - Allow `--version`, `--help`, `-Sdescribe` and a cached `-Spath` to run without Java.
-- Skip Clojure tools installation for `--version`, `--help` and `-Sdescribe`.
+- Skip Clojure tools installation when nothing runs: `--version`, `--help`, `-Sdescribe`, and `-Spath`, `-Stree` or `-P` with `-X` or `-T`.
 - Add `*make-classpath-fn*` to allow in-process classpath resolution.
 
 ## 1.12.6.1673